Various bugfixes while testing changes with TDeck
